backend/pyproject.toml
[project]
name = "home-ledger-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic",
    "openpyxl",
    "pytest",
    "httpx",
    "pyyaml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
